Day 21 part 2 lucky guess (with help from reddit)
Best next: https://gist.github.com/expectancyViolation/f44afc498868e18fce3838764d927ee7
Lucky guess <=

+++ day21/solution/include/day21_solution.h
@@ -5,7 +5,7 @@
 class Day21Solution {
 public:
     using Part1ResultType = int;
-    using Part2ResultType = int;
+    using Part2ResultType = long long;
 
     auto part1(std::istream& inputStream) -> Part1ResultType;
     auto part2(std::istream& inputStream) -> Part2ResultType;
